commit Search ticket by user name or category

+++ ticket/src/main/java/edu/polytech/ticket/repository/TicketRepository.java
@@ -1,14 +1,16 @@
 package edu.polytech.ticket.repository;
 
 import edu.polytech.ticket.entity.TicketEntity;
+import edu.polytech.ticket.enums.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface TicketRepository extends JpaRepository<TicketEntity, Integer> {
-    List<TicketEntity> findByProjectId(Long projectId);
+    List<TicketEntity> findByProjectId(Integer projectId);
 
-    List<TicketEntity> findByProjectIdAndAssignedUserId(Long projectId, Integer assignedUserId);
+    List<TicketEntity> findByProjectIdAndAssignedUserId(Integer projectId, Integer assignedUserId);
+
 
 
 }
